[shared] move yaml_utils
