+++ integration_tests/test_suites/celery-k8s-test-suite/tests/test_monitoring.py
@@ -7,10 +7,10 @@ import time
 from dagster._core.storage.dagster_run import DagsterRunStatus
 from dagster._core.test_utils import poll_for_finished_run
 from dagster._utils.merger import merge_dicts
-from dagster._utils.yaml_utils import merge_yamls
 from dagster_k8s.client import DagsterKubernetesClient
 from dagster_k8s.job import get_job_name_from_run_id
 from dagster_k8s_test_infra.integration_utils import image_pull_policy, launch_run_over_graphql
+from dagster_shared.yaml_utils import merge_yamls
 from dagster_test.test_project import get_test_project_environments_path
 
 IS_BUILDKITE = os.getenv("BUILDKITE") is not None
