+++ integration_tests/test_suites/k8s-test-suite/tests/test_integration.py
@@ -11,7 +11,6 @@ from dagster import (
 from dagster._core.storage.dagster_run import DagsterRunStatus
 from dagster._core.storage.tags import DOCKER_IMAGE_TAG
 from dagster._utils.merger import merge_dicts
-from dagster._utils.yaml_utils import load_yaml_from_path, merge_yamls
 from dagster_k8s.client import DagsterKubernetesClient
 from dagster_k8s.test import wait_for_job_and_get_raw_logs
 from dagster_k8s_test_infra.integration_utils import (
@@ -20,6 +19,7 @@ from dagster_k8s_test_infra.integration_utils import (
     launch_run_over_graphql,
     terminate_run_over_graphql,
 )
+from dagster_shared.yaml_utils import load_yaml_from_path, merge_yamls
 from dagster_test.test_project import (
     get_test_project_docker_image,
     get_test_project_environments_path,
