+++ integration_tests/test_suites/k8s-test-suite/tests/test_k8s_monitoring.py
@@ -6,10 +6,10 @@ import pytest
 from dagster._core.storage.dagster_run import DagsterRunStatus
 from dagster._core.test_utils import poll_for_finished_run
 from dagster._utils.merger import merge_dicts
-from dagster._utils.yaml_utils import load_yaml_from_path
 from dagster_k8s.client import DagsterKubernetesClient
 from dagster_k8s.job import get_job_name_from_run_id
 from dagster_k8s_test_infra.integration_utils import image_pull_policy, launch_run_over_graphql
+from dagster_shared.yaml_utils import load_yaml_from_path
 from dagster_test.test_project import get_test_project_environments_path
