+++ python_modules/dagster-graphql/dagster_graphql/schema/instigation.py
@@ -26,7 +26,7 @@ from dagster._core.storage.tags import REPOSITORY_LABEL_TAG, TagType, get_tag_ty
 from dagster._core.utils import is_valid_run_id
 from dagster._core.workspace.permissions import Permissions
 from dagster._utils.error import SerializableErrorInfo, serializable_error_info_from_exc_info
-from dagster._utils.yaml_utils import dump_run_config_yaml
+from dagster_shared.yaml_utils import dump_run_config_yaml
 
 from dagster_graphql.implementation.fetch_instigators import get_tick_log_events
 from dagster_graphql.implementation.fetch_schedules import get_schedule_next_tick
