+++ python_modules/dagster-graphql/dagster_graphql/schema/runs.py
@@ -7,7 +7,7 @@ import graphene
 import yaml
 from dagster._core.storage.dagster_run import RunsFilter
 from dagster._utils.error import serializable_error_info_from_exc_info
-from dagster._utils.yaml_utils import load_run_config_yaml
+from dagster_shared.yaml_utils import load_run_config_yaml
 from graphene.types.generic import GenericScalar
 
 from dagster_graphql.implementation.fetch_runs import get_run_ids, get_runs, get_runs_count
