+++ python_modules/dagster/dagster/_core/instance/config.py
@@ -4,6 +4,8 @@ from collections.abc import Mapping, Sequence
 from enum import Enum
 from typing import TYPE_CHECKING, Any, Optional, cast
 
+from dagster_shared.yaml_utils import load_yaml_from_globs
+
 from dagster import (
     Array,
     Bool,
@@ -28,7 +30,6 @@ from dagster._record import record
 from dagster._serdes import class_from_code_pointer
 from dagster._utils.concurrency import get_max_concurrency_limit_value
 from dagster._utils.merger import merge_dicts
-from dagster._utils.yaml_utils import load_yaml_from_globs
 
 if TYPE_CHECKING:
     from dagster._core.definitions.run_request import InstigatorType
