+++ python_modules/dagster/dagster/_core/snap/snap_to_yaml.py
@@ -2,8 +2,9 @@ import json
 from collections.abc import Mapping
 from typing import Any, Optional
 
+from dagster_shared.yaml_utils import dump_run_config_yaml
+
 from dagster._config.snap import ConfigSchemaSnapshot, ConfigTypeSnap
-from dagster._utils.yaml_utils import dump_run_config_yaml
 
 
 def _safe_json_loads(json_str: Optional[str]) -> object:
