+++ python_modules/dagster/dagster_tests/utils_tests/test_pydantic_yaml.py
@@ -5,7 +5,7 @@ from dagster._utils.pydantic_yaml import (
     parse_yaml_file_to_pydantic,
     parse_yaml_file_to_pydantic_sequence,
 )
-from dagster._utils.source_position import HasSourcePositionAndKeyPath
+from dagster_shared.yaml_utils.source_position import HasSourcePositionAndKeyPath
 from pydantic import BaseModel, ValidationError
