+++ python_modules/libraries/dagster-celery/dagster_celery/cli.py
@@ -11,7 +11,7 @@ from dagster._config import post_process_config, validate_config
 from dagster._core.errors import DagsterInvalidConfigError
 from dagster._core.instance import DagsterInstance
 from dagster._utils import mkdir_p
-from dagster._utils.yaml_utils import load_yaml_from_path
+from dagster_shared.yaml_utils import load_yaml_from_path
 
 from dagster_celery.executor import CeleryExecutor, celery_executor
 from dagster_celery.make_app import make_app
