+++ python_modules/libraries/dagster-docker/dagster_docker_tests/test_launch_docker.py
@@ -10,12 +10,12 @@ import docker
 import pytest
 from dagster._core.storage.dagster_run import DagsterRunStatus, RunsFilter
 from dagster._core.test_utils import environ, poll_for_finished_run, poll_for_step_start
-from dagster._utils.yaml_utils import merge_yamls
 from dagster_docker.docker_run_launcher import (
     DOCKER_CONTAINER_ID_TAG,
     DOCKER_IMAGE_TAG,
     DockerRunLauncher,
 )
+from dagster_shared.yaml_utils import merge_yamls
 from dagster_test.test_project import (
     ReOriginatedExternalJobForTest,
     find_local_test_image,
