+++ python_modules/libraries/dagster-docker/dagster_docker_tests/test_launcher_and_executor.py
@@ -6,7 +6,7 @@ from dagster._core.definitions.events import AssetKey
 from dagster._core.storage.dagster_run import DagsterRunStatus
 from dagster._core.test_utils import poll_for_finished_run
 from dagster._utils.merger import merge_dicts
-from dagster._utils.yaml_utils import load_yaml_from_path, merge_yamls
+from dagster_shared.yaml_utils import load_yaml_from_path, merge_yamls
 from dagster_test.test_project import (
     ReOriginatedExternalJobForTest,
     find_local_test_image,
